// lib/types/APFMessage.ts
export enum APFMessage {
  DISCONNECT = 1,
  SERVICE_REQUEST = 5,
  SERVICE_ACCEPT = 6,
  USERAUTH_REQUEST = 50,
  USERAUTH_FAILURE = 51,
  USERAUTH_SUCCESS = 52,
  GLOBAL_REQUEST = 80,
  REQUEST_SUCCESS = 81,
  REQUEST_FAILURE = 82,
  CHANNEL_OPEN = 90,
  CHANNEL_OPEN_CONFIRMATION = 91,
  CHANNEL_OPEN_FAILURE = 92,
  CHANNEL_WINDOW_ADJUST = 93,
  CHANNEL_DATA = 94,
  CHANNEL_CLOSE = 97,
  PROTOCOLVERSION = 192,
  KEEPALIVE_REQUEST = 208,
  KEEPALIVE_REPLY = 209,
  KEEPALIVE_OPTIONS_REQUEST = 210,
  KEEPALIVE_OPTIONS_REPLY = 211,

  TCPFORWARD_REQUEST = -1,
  TCPFORWARDCANCEL_REQUEST = -2,
  UDPSENDTO = -3,
}